streamlined attribute checking and setting

+++ pymodal/collection_parent.py
@@ -4,19 +4,23 @@ import h5py
 from pathlib import Path
 from multiprocessing import Pool, cpu_count
 import pint
+import os
+import inspect
+
+os.environ["HDF5_USE_FILE_LOCKING"] = "FALSE"
+
 
 def save_array(array_info):
     array, dataset_name, file_name = array_info
     with h5py.File(file_name, "a") as hf:
         hf.create_dataset(dataset_name, data=array)
 
-def process_strings(args):
-    strings, start_index, end_index = args
+
+def add_suffix(strings):
     counter = {}
     result = []
 
-    for i in range(start_index, end_index):
-        string = strings[i]
+    for string in strings:
         if string in counter:
             count = counter[string]
             new_string = f"{string}_{count}"
@@ -28,8 +32,23 @@ def process_strings(args):
 
     return result
 
+
+def get_attributes(obj, exclude_list=None):
+    if exclude_list is None:
+        exclude_list = []
+
+    attributes = []
+    for name, value in inspect.getmembers(obj):
+        if (
+            not name.startswith("__")
+            and not inspect.ismethod(value)
+            and name not in exclude_list
+        ):
+            attributes.append(name)
+    return attributes
+
+
 num_processes = cpu_count()
-attributes_to_match = ["measurements_units", "method", "dof", "orientations", "coordinates", "space_units", "domain_start", "domain_end", "domain_span", "domain_resolution", "domain_array", "samples"]
 
 # Check if specified attributes match
 def attributes_match(instance1, instance2, attributes_to_match):
@@ -48,60 +67,63 @@ def attributes_match(instance1, instance2, attributes_to_match):
 
     return True
 
+
 def worker(pair):
-    instance1, instance2 = pair
+    instance1, instance2, attributes_to_match = pair
     return attributes_match(instance1, instance2, attributes_to_match)
 
-def parallel_attributes_match(instances):
+
+def parallel_attributes_match(instances, attributes_to_match):
     first_instance = instances[0]
     remaining_instances = instances[1:]
 
     with Pool(num_processes) as pool:
-        results = pool.map(worker, [(first_instance, instance) for instance in remaining_instances])
+        results = pool.map(
+            worker,
+            [
+                (first_instance, instance, attributes_to_match)
+                for instance in remaining_instances
+            ],
+        )
 
     if np.all(results):
         return True
     return False
 
-class _collection():
+
+class _collection:
     def __init__(self, exp_list: list[_signal], path: Path = Path("temp.h5")):
 
         self.path = Path(path)
         if self.path.exists():
             self.path.unlink()
 
-        self.label = list([exp.label for exp in exp_list])
+        self.label = add_suffix(list([exp.label for exp in exp_list]))
 
-        string_count = len(self.label)
-        chunk_size = string_count // num_processes
-        start_indices = [i * chunk_size for i in range(num_processes)]
-        end_indices = start_indices[1:] + [string_count]
-
-        pool = Pool(processes=num_processes)
-        args = [(self.label, start, end) for start, end in zip(start_indices, end_indices)]
-        results = pool.map(process_strings, args)
-        pool.close()
-        pool.join()
-
-        self.label = list([string for sublist in results for string in sublist])
-        
-        assert parallel_attributes_match(exp_list)
-        
-        array_info = [(array, f"{self.label[i]}", self.path) for i, array in enumerate([exp.measurements for exp in exp_list])]
+        attributes_to_match = get_attributes(exp_list[0], ["measurements", "label"])
+        assert parallel_attributes_match(exp_list, attributes_to_match)
+        for attribute in attributes_to_match:
+            setattr(self, attribute, getattr(exp_list[0], attribute))
+        array_info = [
+            (array.magnitude, f"{self.label[i]}", self.path)
+            for i, array in enumerate([exp.measurements for exp in exp_list])
+        ]
         with Pool(num_processes) as pool:
             pool.map(save_array, array_info)
-        pool.close()
-        pool.join()
+        del exp_list
         self.file = h5py.File(self.path, "r")
+        print("Keys: %s" % self.file.keys())
         self.measurements = list([self.file[f"{label}"] for label in self.label])
 
-    def close(self):
+    def close(self, keep: bool = False):
         self.file.close()
-        self.path.unlink()
+        if not keep:
+            self.path.unlink()
+
 
 if __name__ == "__main__":
     from pymodal import frf
-    
+
     time = np.arange(0, 30 + 0.05, 0.1)
     signal = np.sin(1 * time)
     signal = np.vstack((signal, np.sin(2 * time)))
@@ -112,9 +134,9 @@ if __name__ == "__main__":
     signal = np.fft.fft(signal, axis=0)
     signal_1 = signal + 1
     signal_2 = signal + 2
-    test_object_0 = frf(signal, freq_end=5)
-    test_object_1 = frf(signal_1, freq_end=5)
-    test_object_2 = frf(signal_2, freq_end=5)
+    test_object_0 = _signal(signal, domain_end=5)
+    test_object_1 = _signal(signal_1, domain_end=5)
+    test_object_2 = _signal(signal_2, domain_end=5)
     test_collection = _collection([test_object_0, test_object_1, test_object_2])
     print(test_collection.measurements[0])
     test_collection.close()
